pyproject.toml
[project]
name = "py-api-test-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "smoke: marks tests as smoke tests (deselect with '-m \"not smoke\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "security: marks tests as security tests (deselect with '-m \"not security\"')",
]
